♻️ refactor: change Icon component name to ColorBox and update Dashboard header
