+++ src/@shared/index.ts
@@ -6,7 +6,7 @@ export { Header } from "./Header";
 export { SelectBox } from "./SelectBox";
 export { TextField } from "./TextField";
 export { DashboardHeader } from "./DashboardHeader";
-export { Icon } from "./Icon";
+export { ColorBox } from "./ColorBox";
 export { TypeIcon } from "./TypeIcon";
 export { Modal } from "./Modal";
